+++ inc/schedule.h
@@ -36,14 +36,12 @@ struct task_struct {
 };
 
 long privilege_task_create(void(*func)());
-void do_exec(void(*func)());
-void sched_init(void);
+void idle_task();
 void schedule(void);
 void timer_tick(void);
 void context_switch(struct task_struct* next);
 void cpu_switch_to(struct task_struct* prev, struct task_struct* next);
 void task_preemption();
-void wait(long);
 long do_get_taskid();
 void do_exec(void(*func)());
 void do_fork();
