+++ src/schedule.c
@@ -3,9 +3,7 @@
 #include "exce.h"
 #include "mm.h"
 #include "printf.h"
-#include "timer.h"
 #include "uart.h"
-#include "svc.h"
 #include "sysregs.h"
 #include "string.h"
 
@@ -13,96 +11,21 @@
     get ksp base vs get ksp top
     task manage seperate task_struct from ksp
     current -> tpidr_el1
-    init _bss error??
 */
 
-struct task_struct *current;                   //currently executing task
-struct task_struct *task_pool[NR_TASKS]={0};   //task array
+static struct task_struct init_task;
+struct task_struct *task_pool[NR_TASKS]={&init_task, 0};   //task array
+struct task_struct *current = &init_task;      //currently executing task
 
-void delay(int cnt)
-{
-    while(cnt--);
-}
 
-void foo(){// user code
-    int tmp = 5;
-    printf("Task %d after  exec, tmp address 0x%x, tmp value %d\n", sys_get_taskid(), &tmp, tmp);
-    sys_exit(0);
-}
-
-void test() {//user code
-    int cnt = 1;
-    int pid;
-    #ifdef __DEBUG
-    printf("[test] Task_id: %d, cnt: %d\n", current->task_id, cnt);
-    #endif//__DEBUG
-    if ((pid = sys_fork()) == 0) {
-        sys_fork();
-        delay(100000);
-        sys_fork();
+void idle_task(){
+    while(1){
         #ifdef __DEBUG
-        printf("[tast] Task_id: %d, cnt: %d\n", current->task_id, cnt);
-        #endif//__DEBUG
-        while(cnt < 10){
-            printf("Task id: %d, cnt: %d\n", sys_get_taskid(), cnt);
-            delay(100000);
-            ++cnt;
-        }
-        sys_exit(0);
-    printf("Should not be printed\n");
-    } else {
-        printf("Task %d before exec, cnt address 0x%x, cnt value %d\n", sys_get_taskid(), &cnt, cnt);
-        wait(pid);
-        sys_exec(foo);
-  }
-}
-
-void user_test(){//kernel code
-    do_exec(test);
-}
-
-void startup_task()
-{
-    // for(int i = 0; i < 5; ++i) { // N should > 2
-    //     privilege_task_create(foo);
-    // }
-    long pid = privilege_task_create(user_test);
-    if (pid < 0)
-        uart_puts("error while starting user_test\n");
-    core_timer_enable();
-
-    while (1){ // idle here
-        #ifdef __DEBUG
-        delay(100000000);
+        int cnt = 100000000;
+        while(cnt--);
         #endif//DEBUG
         schedule();
-    }
-}
-
-void _load_contex(){
-    asm ("mov x10, #0");
-    asm ("add x8, x0, x10");
-    asm ("ldp x19, x20, [x8], #16");
-    asm ("ldp x21, x22, [x8], #16");
-    asm ("ldp x23, x24, [x8], #16");
-    asm ("ldp x25, x26, [x8], #16");
-    asm ("ldp x27, x28, [x8], #16");
-    asm ("ldp x29, x30, [x8], #16");
-    asm ("ldr x9, [x8]");
-    asm ("mov sp, x9");
-}
-
-void sched_init(void)
-{
-    for(int i=0; i<64;i++){
-        task_pool[i]=0;
-    }
-    long pid = privilege_task_create(&startup_task);
-    if (pid < 0)
-        uart_puts("error while starting idle_task\n");
-    current = task_pool[0];
-    _load_contex(current);
-    uart_puts("[sched_init] should not got here!\n");
+  }
 }
 
 void schedule(void)
@@ -138,8 +61,8 @@ void _context_switch()
     asm ("ldp x29, x30, [x8], #16");
     asm ("ldr x9, [x8]");
     asm ("mov sp, x9");
-	// msr tpidr_el1, x1         // store thread identifying information, for OS management purposes.
-    return;                      // retruen to $lr($x30)
+	// asm ("msr tpidr_el1, x1");     // store thread identifying information, for OS management purposes.
+    return;                        // retruen to $lr($x30)
 }
 
 void context_switch(struct task_struct *next) 
@@ -186,21 +109,6 @@ long privilege_task_create(void(*func)())
     printf("[pcreate] Task_id: %d\tpTask: 0x%X\tksp: 0x%X\n", task_id, pTask, pTask->cpu_context.sp);
     #endif//__DEBUG
     return task_id;
-}
-
-void wait(long task_id){
-    #ifdef __DEBUG
-    printf("[wait] Task_id: %d\n", current->task_id);
-    #endif//__DEBUG
-    while (1)
-    {
-        if(task_pool[task_id]->state == TASK_ZOMBIE){
-            task_pool[task_id] = 0;
-            break;
-        }
-        sys_sched_yield();
-    }
-    
 }
 
 long do_get_taskid(){
